chuyển hướng Guest mode Profile customer: thêm chức năng vé của tôi + ưu đãi của tôi Thêm popup poster khi đăng nhập

+++ app/src/main/java/com/example/cinemabookingapp/core/base/AuthActivity.java
@@ -16,7 +16,7 @@ public class AuthActivity extends BaseActivity{
         authStateListener = authState -> {
             FirebaseUser user = authState.getCurrentUser();
             ServiceProvider.getInstance().getAuthenticationService().getCurrentAuthUser();
-            if (user == null) {
+            if (user == null && !isFinishing()) {
                 ServiceProvider.getInstance().getAuthenticationService().removeCurrentAuthUser();
                 // User signed out, redirect to Login
                 AppNavigator.goToLogin(this);
